requirements.txt
streamlit
python-dotenv
langchain-core
langchain-community
langserve
fastapi
uvicorn[standard]
sse_starlette